requirements.txt
faiss_cpu==1.9.0
faiss_gpu==1.7.2
langgraph==0.4.5
numpy==2.2.6
openai==1.81.0
python-dotenv==1.1.0
sympy==1.13.1
